refactor(kernel/memory/physical): update structs field types

+++ kernel/src/memory/physical/buddy_allocator.rs
@@ -6,6 +6,7 @@ use super::PAGE_SIZE;
 
 static mut BUDDY_ALLOCATOR: BuddyAllocator = BuddyAllocator::null();
 
+#[repr(C, packed)]
 struct PageInfo {
     /// - True: Free
     /// - False: Used
@@ -25,10 +26,11 @@ impl PageInfo {
     }
 }
 
+#[repr(C, packed)]
 pub struct BuddyAllocator {
-    page_count: usize,
+    page_count: u32,
 
-    max_order: usize,
+    max_order: u8,
 
     page_info: &'static mut [PageInfo],
 
@@ -46,7 +48,7 @@ impl BuddyAllocator {
 
     pub fn pre_init(memory_size: usize) -> usize {
         unsafe {
-            BUDDY_ALLOCATOR.page_count = (memory_size + PAGE_SIZE - 1) / PAGE_SIZE;
+            BUDDY_ALLOCATOR.page_count = ((memory_size + PAGE_SIZE - 1) / PAGE_SIZE) as u32;
             BUDDY_ALLOCATOR.max_order = {
                 let mut order = 0;
                 let mut pages = BUDDY_ALLOCATOR.page_count;
@@ -56,16 +58,16 @@ impl BuddyAllocator {
                 }
                 order
             };
-            size_of::<PageInfo>() * BUDDY_ALLOCATOR.page_count
-                + size_of::<usize>() * (BUDDY_ALLOCATOR.max_order + 1)
+            size_of::<PageInfo>() * BUDDY_ALLOCATOR.page_count as usize
+                + size_of::<usize>() * (BUDDY_ALLOCATOR.max_order as usize + 1)
         }
     }
 
     pub fn init(addr: u64, size: usize) {
         unsafe {
             BUDDY_ALLOCATOR.page_info =
-                from_raw_parts_mut(addr as *mut PageInfo, BUDDY_ALLOCATOR.page_count);
-            for i in 0..BUDDY_ALLOCATOR.page_count {
+                from_raw_parts_mut(addr as *mut PageInfo, BUDDY_ALLOCATOR.page_count as usize);
+            for i in 0..(BUDDY_ALLOCATOR.page_count as usize) {
                 BUDDY_ALLOCATOR.page_info[i] = PageInfo::null();
                 if i >= (addr as usize) / PAGE_SIZE
                     && i < (addr as usize + size + PAGE_SIZE - 1) / PAGE_SIZE
@@ -77,9 +79,9 @@ impl BuddyAllocator {
             BUDDY_ALLOCATOR.free_list = from_raw_parts_mut(
                 (addr + size_of::<PageInfo>() as u64 * BUDDY_ALLOCATOR.page_count as u64)
                     as *mut usize,
-                BUDDY_ALLOCATOR.max_order + 1,
+                BUDDY_ALLOCATOR.max_order as usize + 1,
             );
-            for i in 0..=BUDDY_ALLOCATOR.max_order {
+            for i in 0..=(BUDDY_ALLOCATOR.max_order as usize) {
                 BUDDY_ALLOCATOR.free_list[i] = 0;
             }
         }
